+++ web3/tools/pytest_ethereum/session.py
@@ -1,0 +1,3 @@
+from sqlalchemy.orm import scoped_session, sessionmaker
+
+Session = scoped_session(sessionmaker())
